Se cambio int por double para el sn. falta arreglar cuando se crea un articulo

+++ Muebleria/CodigoDeBarras.cs
@@ -13,7 +13,7 @@ namespace Muebleria
 
         }
 
-        public string generarCodigoDeBarras(long sn)
+        public string generarCodigoDeBarras(double sn)
         {
             string cdb = "*" + sn.ToString() + "*";
             return cdb;
